Moved the fulltext search adapter back to skeleton and auto-attach it to skeletons
